working version with chaos related to docker

+++ embeddingModel/saveJinaLocally.py
@@ -1,12 +1,21 @@
 from transformers import AutoModel
 import torch
+import os
 
-model = AutoModel.from_pretrained("jinaai/jina-clip-v1", trust_remote_code=True)
-model.eval()
-# Set the device (GPU if available, otherwise CPU)
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-model.to(device)  # Move the model to the device
+model_path = "./jina_clip_v1_model"
 
-model_path = "./api/jina_clip_v1_model"
-model.save_pretrained(model_path, safe_serialization=False)
-torch.save(model, "./api/jina_clip_v1_model/jina.pt")
+# Check if the folder already exists
+if not os.path.exists(model_path):
+    print(f"Loading and saving model to {model_path}")
+    model = AutoModel.from_pretrained("jinaai/jina-clip-v1", trust_remote_code=True)
+    model.eval()
+    # Set the device (GPU if available, otherwise CPU)
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    model.to(device)  # Move the model to the device
+
+    # Save the model
+    model.save_pretrained(model_path, safe_serialization=False)
+    torch.save(model, os.path.join(model_path, "jina.pt"))
+    print("Model saved successfully")
+else:
+    print(f"Model folder {model_path} already exists. Skipping download.")
